@@ -6,32 +6,32 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  freezed_annotation: ^0.12.0
-  floor: ^0.14.0
-  retrofit: ^1.3.4+1
-  json_annotation: ^3.1.0
-  get_it: ^4.0.2
-  injectable: ^1.0.4
-  cached_network_image: ^2.3.2+1
-  auto_route: ^0.6.7
-  flutter_svg: ^0.19.0
-  flutter_bloc: ^6.0.6
+  freezed_annotation: ^0.14.1
+  floor: ^1.0.1
+  retrofit: ^2.0.0-beta1
+  json_annotation: ^4.0.1
+  get_it: ^6.0.0
+  injectable: ^1.2.2
+  cached_network_image: ^3.0.0
+  auto_route: ^1.0.2
+  flutter_svg: ^0.21.0-nullsafety.0
+  flutter_bloc: ^7.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   build_runner: ^1.10.2
-  freezed: ^0.12.1
-  floor_generator: ^0.14.0
-  retrofit_generator: ^1.3.8+5
-  json_serializable: ^3.5.0
-  injectable_generator: ^1.0.4
-  auto_route_generator: ^0.6.8
+  freezed: ^0.14.1+1
+  floor_generator: ^1.0.1
+  retrofit_generator: ^2.0.0-beta3
+  json_serializable: ^4.1.0
+  injectable_generator: ^1.2.2
+  auto_route_generator: ^1.0.2
 
 flutter:
   uses-material-design: true
