@@ -1,6 +1,7 @@
 import 'package:floor/floor.dart';
-import 'package:pokeapp/domain/constants/pokemon_type.dart';
-import 'package:pokeapp/domain/entities/pokemon.dart';
+
+import '../../domain/constants/pokemon_type.dart';
+import '../../domain/entities/pokemon.dart';
 
 @Entity(tableName: "Pokemons")
 class PokemonData {
@@ -9,40 +10,52 @@ class PokemonData {
 
   final String name;
 
+  @ColumnInfo(name: "image_url")
   final String imageUrl;
 
-  int baseExperience;
-  int height;
-  int weight;
-  String types;
+  @ColumnInfo(name: "base_experience")
+  int? baseExperience;
 
-  PokemonData(
-      {this.id,
-      this.name,
-      this.imageUrl,
-      this.baseExperience,
-      this.height,
-      this.weight,
-      this.types});
+  int? height;
+
+  int? weight;
+
+  @ColumnInfo(name: "primary_type")
+  PokemonType? primaryType;
+
+  @ColumnInfo(name: "secondary_type")
+  PokemonType? secondaryType;
+
+  PokemonData({
+    required this.id,
+    required this.name,
+    required this.imageUrl,
+    required this.baseExperience,
+    required this.height,
+    required this.weight,
+    required this.primaryType,
+    required this.secondaryType,
+  });
 
   Pokemon toEntity() => Pokemon(
-      id: id,
-      name: name,
-      imageUrl: imageUrl,
-      baseExperience: baseExperience,
-      height: height,
-      weight: weight,
-      types: types
-          .split(",")
-          .map((type) => PokemonTypeX.fromValue(type))
-          .toList());
+        id: id,
+        name: name,
+        imageUrl: imageUrl,
+        baseExperience: baseExperience,
+        height: height,
+        weight: weight,
+        primaryType: primaryType,
+        secondaryType: secondaryType,
+      );
 
   factory PokemonData.fromEntity(Pokemon entity) => PokemonData(
-      id: entity.id,
-      name: entity.name,
-      imageUrl: entity.imageUrl,
-      baseExperience: entity.baseExperience,
-      height: entity.height,
-      weight: entity.weight,
-      types: entity.types.map((type) => type.value).join(","));
+        id: entity.id,
+        name: entity.name,
+        imageUrl: entity.imageUrl,
+        baseExperience: entity.baseExperience,
+        height: entity.height,
+        weight: entity.weight,
+        primaryType: entity.primaryType,
+        secondaryType: entity.secondaryType,
+      );
 }
