@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'pokemon.dart';
 
@@ -9,46 +9,51 @@ part of 'pokemon.dart';
 
 T _$identity<T>(T value) => value;
 
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 /// @nodoc
 class _$PokemonTearOff {
   const _$PokemonTearOff();
 
-// ignore: unused_element
   _Pokemon call(
-      {@required int id,
-      @required String name,
-      @required String imageUrl,
-      int baseExperience,
-      int height,
-      int weight,
-      List<PokemonType> types}) {
+      {required int id,
+      required String name,
+      required String imageUrl,
+      required PokemonType? primaryType,
+      PokemonType? secondaryType,
+      int? baseExperience,
+      int? height,
+      int? weight}) {
     return _Pokemon(
       id: id,
       name: name,
       imageUrl: imageUrl,
+      primaryType: primaryType,
+      secondaryType: secondaryType,
       baseExperience: baseExperience,
       height: height,
       weight: weight,
-      types: types,
     );
   }
 }
 
 /// @nodoc
-// ignore: unused_element
 const $Pokemon = _$PokemonTearOff();
 
 /// @nodoc
 mixin _$Pokemon {
-  int get id;
-  String get name;
-  String get imageUrl;
-  int get baseExperience;
-  int get height;
-  int get weight;
-  List<PokemonType> get types;
+  int get id => throw _privateConstructorUsedError;
+  String get name => throw _privateConstructorUsedError;
+  String get imageUrl => throw _privateConstructorUsedError;
+  PokemonType? get primaryType => throw _privateConstructorUsedError;
+  PokemonType? get secondaryType => throw _privateConstructorUsedError;
+  int? get baseExperience => throw _privateConstructorUsedError;
+  int? get height => throw _privateConstructorUsedError;
+  int? get weight => throw _privateConstructorUsedError;
 
-  $PokemonCopyWith<Pokemon> get copyWith;
+  @JsonKey(ignore: true)
+  $PokemonCopyWith<Pokemon> get copyWith => throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -59,10 +64,11 @@ abstract class $PokemonCopyWith<$Res> {
       {int id,
       String name,
       String imageUrl,
-      int baseExperience,
-      int height,
-      int weight,
-      List<PokemonType> types});
+      PokemonType? primaryType,
+      PokemonType? secondaryType,
+      int? baseExperience,
+      int? height,
+      int? weight});
 }
 
 /// @nodoc
@@ -75,24 +81,48 @@ class _$PokemonCopyWithImpl<$Res> implements $PokemonCopyWith<$Res> {
 
   @override
   $Res call({
-    Object id = freezed,
-    Object name = freezed,
-    Object imageUrl = freezed,
-    Object baseExperience = freezed,
-    Object height = freezed,
-    Object weight = freezed,
-    Object types = freezed,
+    Object? id = freezed,
+    Object? name = freezed,
+    Object? imageUrl = freezed,
+    Object? primaryType = freezed,
+    Object? secondaryType = freezed,
+    Object? baseExperience = freezed,
+    Object? height = freezed,
+    Object? weight = freezed,
   }) {
     return _then(_value.copyWith(
-      id: id == freezed ? _value.id : id as int,
-      name: name == freezed ? _value.name : name as String,
-      imageUrl: imageUrl == freezed ? _value.imageUrl : imageUrl as String,
+      id: id == freezed
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as int,
+      name: name == freezed
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as String,
+      imageUrl: imageUrl == freezed
+          ? _value.imageUrl
+          : imageUrl // ignore: cast_nullable_to_non_nullable
+              as String,
+      primaryType: primaryType == freezed
+          ? _value.primaryType
+          : primaryType // ignore: cast_nullable_to_non_nullable
+              as PokemonType?,
+      secondaryType: secondaryType == freezed
+          ? _value.secondaryType
+          : secondaryType // ignore: cast_nullable_to_non_nullable
+              as PokemonType?,
       baseExperience: baseExperience == freezed
           ? _value.baseExperience
-          : baseExperience as int,
-      height: height == freezed ? _value.height : height as int,
-      weight: weight == freezed ? _value.weight : weight as int,
-      types: types == freezed ? _value.types : types as List<PokemonType>,
+          : baseExperience // ignore: cast_nullable_to_non_nullable
+              as int?,
+      height: height == freezed
+          ? _value.height
+          : height // ignore: cast_nullable_to_non_nullable
+              as int?,
+      weight: weight == freezed
+          ? _value.weight
+          : weight // ignore: cast_nullable_to_non_nullable
+              as int?,
     ));
   }
 }
@@ -106,10 +136,11 @@ abstract class _$PokemonCopyWith<$Res> implements $PokemonCopyWith<$Res> {
       {int id,
       String name,
       String imageUrl,
-      int baseExperience,
-      int height,
-      int weight,
-      List<PokemonType> types});
+      PokemonType? primaryType,
+      PokemonType? secondaryType,
+      int? baseExperience,
+      int? height,
+      int? weight});
 }
 
 /// @nodoc
@@ -123,24 +154,48 @@ class __$PokemonCopyWithImpl<$Res> extends _$PokemonCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object id = freezed,
-    Object name = freezed,
-    Object imageUrl = freezed,
-    Object baseExperience = freezed,
-    Object height = freezed,
-    Object weight = freezed,
-    Object types = freezed,
+    Object? id = freezed,
+    Object? name = freezed,
+    Object? imageUrl = freezed,
+    Object? primaryType = freezed,
+    Object? secondaryType = freezed,
+    Object? baseExperience = freezed,
+    Object? height = freezed,
+    Object? weight = freezed,
   }) {
     return _then(_Pokemon(
-      id: id == freezed ? _value.id : id as int,
-      name: name == freezed ? _value.name : name as String,
-      imageUrl: imageUrl == freezed ? _value.imageUrl : imageUrl as String,
+      id: id == freezed
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as int,
+      name: name == freezed
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as String,
+      imageUrl: imageUrl == freezed
+          ? _value.imageUrl
+          : imageUrl // ignore: cast_nullable_to_non_nullable
+              as String,
+      primaryType: primaryType == freezed
+          ? _value.primaryType
+          : primaryType // ignore: cast_nullable_to_non_nullable
+              as PokemonType?,
+      secondaryType: secondaryType == freezed
+          ? _value.secondaryType
+          : secondaryType // ignore: cast_nullable_to_non_nullable
+              as PokemonType?,
       baseExperience: baseExperience == freezed
           ? _value.baseExperience
-          : baseExperience as int,
-      height: height == freezed ? _value.height : height as int,
-      weight: weight == freezed ? _value.weight : weight as int,
-      types: types == freezed ? _value.types : types as List<PokemonType>,
+          : baseExperience // ignore: cast_nullable_to_non_nullable
+              as int?,
+      height: height == freezed
+          ? _value.height
+          : height // ignore: cast_nullable_to_non_nullable
+              as int?,
+      weight: weight == freezed
+          ? _value.weight
+          : weight // ignore: cast_nullable_to_non_nullable
+              as int?,
     ));
   }
 }
@@ -148,16 +203,14 @@ class __$PokemonCopyWithImpl<$Res> extends _$PokemonCopyWithImpl<$Res>
 /// @nodoc
 class _$_Pokemon with DiagnosticableTreeMixin implements _Pokemon {
   const _$_Pokemon(
-      {@required this.id,
-      @required this.name,
-      @required this.imageUrl,
+      {required this.id,
+      required this.name,
+      required this.imageUrl,
+      required this.primaryType,
+      this.secondaryType,
       this.baseExperience,
       this.height,
-      this.weight,
-      this.types})
-      : assert(id != null),
-        assert(name != null),
-        assert(imageUrl != null);
+      this.weight});
 
   @override
   final int id;
@@ -166,17 +219,19 @@ class _$_Pokemon with DiagnosticableTreeMixin implements _Pokemon {
   @override
   final String imageUrl;
   @override
-  final int baseExperience;
+  final PokemonType? primaryType;
   @override
-  final int height;
+  final PokemonType? secondaryType;
   @override
-  final int weight;
+  final int? baseExperience;
   @override
-  final List<PokemonType> types;
+  final int? height;
+  @override
+  final int? weight;
 
   @override
   String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
-    return 'Pokemon(id: $id, name: $name, imageUrl: $imageUrl, baseExperience: $baseExperience, height: $height, weight: $weight, types: $types)';
+    return 'Pokemon(id: $id, name: $name, imageUrl: $imageUrl, primaryType: $primaryType, secondaryType: $secondaryType, baseExperience: $baseExperience, height: $height, weight: $weight)';
   }
 
   @override
@@ -187,10 +242,11 @@ class _$_Pokemon with DiagnosticableTreeMixin implements _Pokemon {
       ..add(DiagnosticsProperty('id', id))
       ..add(DiagnosticsProperty('name', name))
       ..add(DiagnosticsProperty('imageUrl', imageUrl))
+      ..add(DiagnosticsProperty('primaryType', primaryType))
+      ..add(DiagnosticsProperty('secondaryType', secondaryType))
       ..add(DiagnosticsProperty('baseExperience', baseExperience))
       ..add(DiagnosticsProperty('height', height))
-      ..add(DiagnosticsProperty('weight', weight))
-      ..add(DiagnosticsProperty('types', types));
+      ..add(DiagnosticsProperty('weight', weight));
   }
 
   @override
@@ -204,15 +260,19 @@ class _$_Pokemon with DiagnosticableTreeMixin implements _Pokemon {
             (identical(other.imageUrl, imageUrl) ||
                 const DeepCollectionEquality()
                     .equals(other.imageUrl, imageUrl)) &&
+            (identical(other.primaryType, primaryType) ||
+                const DeepCollectionEquality()
+                    .equals(other.primaryType, primaryType)) &&
+            (identical(other.secondaryType, secondaryType) ||
+                const DeepCollectionEquality()
+                    .equals(other.secondaryType, secondaryType)) &&
             (identical(other.baseExperience, baseExperience) ||
                 const DeepCollectionEquality()
                     .equals(other.baseExperience, baseExperience)) &&
             (identical(other.height, height) ||
                 const DeepCollectionEquality().equals(other.height, height)) &&
             (identical(other.weight, weight) ||
-                const DeepCollectionEquality().equals(other.weight, weight)) &&
-            (identical(other.types, types) ||
-                const DeepCollectionEquality().equals(other.types, types)));
+                const DeepCollectionEquality().equals(other.weight, weight)));
   }
 
   @override
@@ -221,11 +281,13 @@ class _$_Pokemon with DiagnosticableTreeMixin implements _Pokemon {
       const DeepCollectionEquality().hash(id) ^
       const DeepCollectionEquality().hash(name) ^
       const DeepCollectionEquality().hash(imageUrl) ^
+      const DeepCollectionEquality().hash(primaryType) ^
+      const DeepCollectionEquality().hash(secondaryType) ^
       const DeepCollectionEquality().hash(baseExperience) ^
       const DeepCollectionEquality().hash(height) ^
-      const DeepCollectionEquality().hash(weight) ^
-      const DeepCollectionEquality().hash(types);
+      const DeepCollectionEquality().hash(weight);
 
+  @JsonKey(ignore: true)
   @override
   _$PokemonCopyWith<_Pokemon> get copyWith =>
       __$PokemonCopyWithImpl<_Pokemon>(this, _$identity);
@@ -233,28 +295,33 @@ class _$_Pokemon with DiagnosticableTreeMixin implements _Pokemon {
 
 abstract class _Pokemon implements Pokemon {
   const factory _Pokemon(
-      {@required int id,
-      @required String name,
-      @required String imageUrl,
-      int baseExperience,
-      int height,
-      int weight,
-      List<PokemonType> types}) = _$_Pokemon;
+      {required int id,
+      required String name,
+      required String imageUrl,
+      required PokemonType? primaryType,
+      PokemonType? secondaryType,
+      int? baseExperience,
+      int? height,
+      int? weight}) = _$_Pokemon;
 
   @override
-  int get id;
+  int get id => throw _privateConstructorUsedError;
   @override
-  String get name;
+  String get name => throw _privateConstructorUsedError;
   @override
-  String get imageUrl;
+  String get imageUrl => throw _privateConstructorUsedError;
   @override
-  int get baseExperience;
+  PokemonType? get primaryType => throw _privateConstructorUsedError;
   @override
-  int get height;
+  PokemonType? get secondaryType => throw _privateConstructorUsedError;
   @override
-  int get weight;
+  int? get baseExperience => throw _privateConstructorUsedError;
   @override
-  List<PokemonType> get types;
+  int? get height => throw _privateConstructorUsedError;
   @override
-  _$PokemonCopyWith<_Pokemon> get copyWith;
+  int? get weight => throw _privateConstructorUsedError;
+  @override
+  @JsonKey(ignore: true)
+  _$PokemonCopyWith<_Pokemon> get copyWith =>
+      throw _privateConstructorUsedError;
 }
