@@ -1,6 +1,6 @@
-import 'package:auto_route/auto_route_annotations.dart';
-import 'package:pokeapp/app/details/pokemon_details_page.dart';
+import 'package:auto_route/annotations.dart';
 
+import 'details/pokemon_details_page.dart';
 import 'list/pokemon_list_page.dart';
 
 @MaterialAutoRouter(
