@@ -4,64 +4,61 @@
 // AutoRouteGenerator
 // **************************************************************************
 
-// ignore_for_file: public_member_api_docs
+import 'package:auto_route/auto_route.dart' as _i1;
+import 'package:flutter/material.dart' as _i5;
 
-import 'package:auto_route/auto_route.dart';
-import 'package:flutter/material.dart';
+import '../domain/entities/pokemon.dart' as _i4;
+import 'details/pokemon_details_page.dart' as _i3;
+import 'list/pokemon_list_page.dart' as _i2;
 
-import '../domain/entities/pokemon.dart';
-import 'details/pokemon_details_page.dart';
-import 'list/pokemon_list_page.dart';
+class Router extends _i1.RootStackRouter {
+  Router();
 
-class Routes {
-  static const String pokemonListPage = '/';
-  static const String pokemonDetailsPage = '/pokemon-details-page';
-  static const all = <String>{
-    pokemonListPage,
-    pokemonDetailsPage,
+  @override
+  final Map<String, _i1.PageFactory> pagesMap = {
+    PokemonListPageRoute.name: (entry) {
+      return _i1.MaterialPageX(entry: entry, child: _i2.PokemonListPage());
+    },
+    PokemonDetailsPageRoute.name: (entry) {
+      var args = entry.routeData.argsAs<PokemonDetailsPageRouteArgs>();
+      return _i1.MaterialPageX(
+          entry: entry,
+          child: _i3.PokemonDetailsPage(
+              pokemon: args.pokemon, backgroundColor: args.backgroundColor));
+    }
   };
+
+  @override
+  List<_i1.RouteConfig> get routes => [
+        _i1.RouteConfig(PokemonListPageRoute.name, path: '/'),
+        _i1.RouteConfig(PokemonDetailsPageRoute.name,
+            path: '/pokemon-details-page')
+      ];
 }
 
-class Router extends RouterBase {
-  @override
-  List<RouteDef> get routes => _routes;
-  final _routes = <RouteDef>[
-    RouteDef(Routes.pokemonListPage, page: PokemonListPage),
-    RouteDef(Routes.pokemonDetailsPage, page: PokemonDetailsPage),
-  ];
-  @override
-  Map<Type, AutoRouteFactory> get pagesMap => _pagesMap;
-  final _pagesMap = <Type, AutoRouteFactory>{
-    PokemonListPage: (data) {
-      return MaterialPageRoute<dynamic>(
-        builder: (context) => PokemonListPage(),
-        settings: data,
-      );
-    },
-    PokemonDetailsPage: (data) {
-      final args = data.getArgs<PokemonDetailsPageArguments>(
-        orElse: () => PokemonDetailsPageArguments(),
-      );
-      return MaterialPageRoute<dynamic>(
-        builder: (context) => PokemonDetailsPage(
-          key: args.key,
-          pokemon: args.pokemon,
-          backgroundColor: args.backgroundColor,
-        ),
-        settings: data,
-      );
-    },
-  };
+class PokemonListPageRoute extends _i1.PageRouteInfo {
+  const PokemonListPageRoute() : super(name, path: '/');
+
+  static const String name = 'PokemonListPageRoute';
 }
 
-/// ************************************************************************
-/// Arguments holder classes
-/// *************************************************************************
+class PokemonDetailsPageRoute
+    extends _i1.PageRouteInfo<PokemonDetailsPageRouteArgs> {
+  PokemonDetailsPageRoute(
+      {required _i4.Pokemon pokemon, required _i5.Color backgroundColor})
+      : super(name,
+            path: '/pokemon-details-page',
+            args: PokemonDetailsPageRouteArgs(
+                pokemon: pokemon, backgroundColor: backgroundColor));
 
-/// PokemonDetailsPage arguments holder class
-class PokemonDetailsPageArguments {
-  final Key key;
-  final Pokemon pokemon;
-  final Color backgroundColor;
-  PokemonDetailsPageArguments({this.key, this.pokemon, this.backgroundColor});
+  static const String name = 'PokemonDetailsPageRoute';
+}
+
+class PokemonDetailsPageRouteArgs {
+  const PokemonDetailsPageRouteArgs(
+      {required this.pokemon, required this.backgroundColor});
+
+  final _i4.Pokemon pokemon;
+
+  final _i5.Color backgroundColor;
 }
