@@ -1,5 +1,5 @@
-import 'package:pokeapp/domain/entities/pokemon.dart';
+import '../entities/pokemon.dart';
 
 abstract class PokemonRepository {
-  Future<List<Pokemon>> getAllPokemons({int lastId});
+  Future<List<Pokemon>> getAllPokemons({int? lastId});
 }
