@@ -1,6 +1,7 @@
 import 'package:injectable/injectable.dart';
-import 'package:pokeapp/domain/entities/pokemon.dart';
-import 'package:pokeapp/domain/repositories/pokemon_repository.dart';
+
+import '../entities/pokemon.dart';
+import '../repositories/pokemon_repository.dart';
 
 @injectable
 class GetAllPokemons {
@@ -8,6 +9,6 @@ class GetAllPokemons {
 
   GetAllPokemons(this._repository);
 
-  Future<List<Pokemon>> call({int lastId}) =>
+  Future<List<Pokemon>> call({int? lastId}) =>
       _repository.getAllPokemons(lastId: lastId);
 }
